feat : added auto task
